refactor(author): modify `Author` types
