+++ packages/readr/components/post/post-credit.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components'
 
-import type { Author } from '~/graphql/query/author'
+import type { Author } from '~/graphql/fragments/author'
 import type { PostDetail } from '~/graphql/query/post'
 
 import MediaLinkList from '../shared/media-link'
