+++ packages/readr/pages/author/[id].tsx
@@ -10,8 +10,8 @@ import client from '~/apollo-client'
 import LayoutGeneral from '~/components/layout/layout-general'
 import ArticleListCard from '~/components/shared/article-list-card'
 import SectionHeading from '~/components/shared/section-heading'
+import type { Author } from '~/graphql/fragments/author'
 import type { Post } from '~/graphql/fragments/post'
-import type { Author } from '~/graphql/query/author'
 import { author as authorQuery } from '~/graphql/query/author'
 import { authorPosts as authorPostsQuery } from '~/graphql/query/post'
 import useInfiniteScroll from '~/hooks/useInfiniteScroll'
